requirements.txt
albumentations==1.3.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
streamlit==1.28.0
tqdm==4.66.1
